Add errors and extract out the forms
