+++ frontend/components/FormError/FormError.tsx
@@ -2,7 +2,7 @@ import { ApolloError } from '@apollo/client'
 import theme from '../../theme'
 
 type ErrorProps = {
-  error: string
+  error: React.ReactNode
 }
 
 const ErrorMessage: React.FC<ErrorProps> = ({ error }) => (
@@ -10,9 +10,18 @@ const ErrorMessage: React.FC<ErrorProps> = ({ error }) => (
     <p className="form-error-message">{error}</p>
 
     <style jsx>{`
+      @keyframes fadeIn {
+        from {
+          opacity: 0;
+        }
+        to {
+          opacity: 1;
+        }
+      }
+
       .form-error {
-        margin: 12px 0;
         border-left: 5px solid ${theme.colors.red};
+        animation: fadeIn 150ms linear;
       }
 
       .form-error-message {
@@ -25,13 +34,16 @@ const ErrorMessage: React.FC<ErrorProps> = ({ error }) => (
 )
 
 type Props = {
-  error?: ApolloError
+  error: ApolloError | string
 }
 
 // Type must be augmented because of https://github.com/apollographql/apollo-link/issues/536
 type NetworkError = ApolloError & { result: { errors: Error[] } }
 
 const FormError: React.FC<Props> = ({ error }) => {
+  if (typeof error === 'string') {
+    return <ErrorMessage error={error} />
+  }
   if (!error || !error.message) return null
 
   const result = (error.networkError as NetworkError)?.result
